requirements.txt
# Bobby's Workshop - Secret Rooms Backend Dependencies

# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Audio processing
pyaudio>=0.2.14
soundfile>=0.12.1
numpy>=1.24.0
scipy>=1.11.0

# Transcription
openai-whisper>=20231117
faster-whisper>=0.10.0  # Faster local Whisper implementation
webrtcvad>=2.0.10  # Voice activity detection

# Audio processing (advanced)
librosa>=0.10.0  # Audio analysis and processing
noisereduce>=3.0.0  # Noise reduction
yt-dlp>=2023.12.30  # URL audio/video extraction

# Speaker diarization
# pyannote.audio>=3.1.0  # For speaker diarization (requires HuggingFace token)
# Note: Uncomment and configure HuggingFace token for pyannote.audio

# Image processing
Pillow>=10.0.0

# USB/Hardware
pyusb>=1.2.1

# Optional: Advanced features
# deepfilternet>=0.5.0  # For advanced noise reduction (Tier 2 feature)
